US010_TC001_Jair and Update the log in
