--- src/test/java/com/libraryAutomation/stepDefinitions/Login_StepDefinitions1.java
@@ -1,55 +1,0 @@
-package com.libraryAutomation.stepDefinitions;
-
-import com.libraryAutomation.pages.LoginPage;
-import com.libraryAutomation.utilities.ConfigurationReader;
-import com.libraryAutomation.utilities.Driver;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-
-public class Login_StepDefinitions1 {
-
-    LoginPage loginPage = new LoginPage();
-
-    @Given("I am on the login page")
-    public void i_am_on_the_login_page() {
-        Driver.getDriver().get(ConfigurationReader.getProperty("url"));
-    }
-
-    @Given("I login as a librarian1")
-    public void i_login_as_a_librarian1() {
-        Driver.getDriver().get(ConfigurationReader.getProperty("url"));
-        loginPage.loginAsLibrary();
-
-    }
-    @Then("dashboard should be displayed")
-    public void dashboard_should_be_displayed() {
-
-    }
-
-    @Given("I login as a student1")
-    public void i_login_as_a_student1() {
-        Driver.getDriver().get(ConfigurationReader.getProperty("url"));
-        loginPage.loginAsStudent();
-
-    }
-
-
-    @When("I enter username {string}")
-    public void i_enter_username(String string) {
-      loginPage.usernameInput.sendKeys(string);
-    }
-    @When("I enter password {string}")
-    public void i_enter_password(String string) {
-        loginPage.passwordInput.sendKeys(string);
-    }
-    @When("I click the sign in button")
-    public void i_click_the_sign_in_button() {
-        loginPage.signInButton.click();
-
-    }
-
-
-
-}
